Move models into demo folder
ConsumptionSaving, cstwMPC, cAndCwithStickyE, FashionVictim, and
SolvingMicroDSOPs are not part of the core library; they are users of
it, and so should live in the Demos folder.

For the sake of reviewability, this commit does not change the contents
of any of the moved files, and so they do not work in this commit; the
import paths will be updated in a subsequent commit.