Fix demo import paths following move
The previous commit moved the demo models from subdirectories of the main
HARK framework to subdirectories of the "Demos" directory, and for the
sake of reviewability, left them in a non-working state.

This commit updates the relative import paths to refer to the new
relative location.

Broadly, old directory structure looked like:

    ./HARK/Model/Code.py
    ./HARK/HARKcore.py

The various models added "../" to the import path, so that `Code.py`
could import `HARKcore.py`.

The new directory structure looks more like:

    ./Demos/Model/Code.py
    ./HARK/HARKcore.py

So, to allow `Code.py` to import `HARKcore.py`, we need to append a
"../HARK/" to the old "../" path.