Rename HARKcore to HARK.core
Rename the core module, and refer to it by its new name in all the code
in the repo that currently refers to it. There are several pieces of
documentation that still refer to HARKcore instead of HARK.core; these
will need to be cleaned up later.

Remove the remaining import path references to the HARK directory, as
now everything imports by the desired module path.